AP: Configure unsolicited broadcast Probe Response interval
Currently unsolicited broadcast Probe Response interval is not configured hence SAP is not sending Un-solicited Probe Response within an interval of 20 TU + 3 TUs.

Configured the unsolicited broadcast probe response interval as 20 TU for the 6 GHz band and 11ax mode. For non-zero interval, the AP schedules unsolicited broadcast Probe Response frames to be transmitted for in-band discovery.

Change-Id: I725b42080f705ef329922f01e5cd65596e7576a6